server: remove the conflict between DB updates and running queries
Each database handle has a shared lock that queries hold while they run, and
that updating the state of the handle with a new database, label, and named
queries acquire exclusively. This meant that handing off a new database would
block for a long-running query.

I dealt with this by setting a timeout on user queries. That works, but can be
annoying when users have a complex query to run that may take a while.

To address this, change the way updates are done:

- If an update arrives and the handle is not locked, it is applied immediately
  (as before).

- But if the handle is locked (busy), record the intended update in an atomic
  field.

- Each time a query exits, it checks whether there is a pending update and
  tries to apply it.

- Each time the UI browses the handles, it checks whether there are pending
  updates and tries to apply them.

If an update cannot be applied immediately, then one or more running queries
hold the lock that prevents it. Either the last such query, or a subsequent UI
visit, will notice the pending query and apply it.