server: fix a nil indirection when updating databases
When checking for an update, don't assume the previous value was non-nil, since
usually it WILL be nil.

I tested this manually by hacking the condition to force it into the right
state, but it's tricky to get this in a unit test without drilling long holes
into the plumbing.