server: add support for fully-programmatic data sources (#24)
Allow a caller to hook up databases that do not require use of the database/sql
package directly.

The substance of the change is to add Queryable and RowSet interfaces, and to
rework the query plumbing to use them. Existing use of *sql.DB is shimmed with
a wrapper implementing Queryable.

I wanted to use a generic interface so that a shim would not be needed, but
this turned out to be more trouble than it was worth, since it infects the rest of
the package with type parameters. It was (much) simpler to use the wrapper.