server: restore separate connections for the local state db (#29)
In #25 I switched away from using a transaction for queries on the local state
database, because it did not interact properly with the queryable interface we
added in #24. But because I'd switched away from having a separate read-only
connection for these queries, this had the undesirable side-effect of allowing
write queries on the local state database.

Switch back to using a separate read-only connection for local state.  Morally
this is a partial revert of #24, but done in a slightly different way.